Fix customer address

+++ app/Actions/CustomerAddress/DeleteCustomerAddressAction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Actions\CustomerAddress;
 
+use App\Exceptions\CannotDeleteDefaultAddressException;
 use App\Models\Customer;
 use App\Models\CustomerAddress;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
@@ -22,6 +23,10 @@ final readonly class DeleteCustomerAddressAction
 
             if ($address->customer_id !== $customer->id) {
                 throw new ModelNotFoundException()->setModel(CustomerAddress::class, $address->id);
+            }
+
+            if ($address->default) {
+                throw CannotDeleteDefaultAddressException::create();
             }
 
             $address->delete();
